% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gassco_outages.R
\name{gassco_outages_parse}
\alias{gassco_outages_parse}
\title{Download and parse the gassco umm.}
\usage{
gassco_outages_parse(path = "https://umm.gassco.no/atom.xml")
}
\arguments{
\item{path}{Path.}
}
\description{
Function returns NULL if there are no messages to parse.
}
\examples{

library(gasscor)
gassco_outages_parse()

}
